Split jtag and alternative interface functions
Flexiband 1 is only able to use the jtag
interface to flash it's fpga. With newer
versions like the flexiband2 and forward
an alternative interface was introduced.
This change splits the two distinct
procedures into two independend functions.
It also removes the properitary naming
of the variables. THis way the functions
are much easier to maintain and to
extend.

Change-Id: I13f0e0c86e77f62e46b038a9c91016eeaf5a2af3